+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Features/Improvements
 * **[Mission Generation]** Given a CAS flight was planned, delay ground force attack until first CAS flight is on station
+* **[Mission Generation]** Add option to switch ATFLIR to LITENING automatically for ground based F-18C flights
 
 ## Fixes
 * **[UI]** Removed deprecated options